Status effects, equipment and traits
Traits can be activated without using the turn and have a cooldown.
Equipment takes the place of traits from the previous the old game.
Status Effects are the same as in the old game. Changed the ui, so it
will display the stats of the character you highlight. Renamed engine
folder.